Separate Obs from Adjoint and adapt for better testing
