+++ pennylane_lightning_gpu/src/algorithms/CMakeLists.txt
@@ -1,7 +1,7 @@
 project(lightning_gpu_algorithms LANGUAGES CXX)
 set(CMAKE_CXX_STANDARD 20)
 
-set(GPU_ALGORITHM_FILES AdjointDiffGPU.hpp AdjointDiffGPU.cpp CACHE INTERNAL "" FORCE)
+set(GPU_ALGORITHM_FILES AdjointDiffGPU.hpp AdjointDiffGPU.cpp ObservablesGPU.hpp CACHE INTERNAL "" FORCE)
 add_library(lightning_gpu_algorithms STATIC ${GPU_ALGORITHM_FILES})
 
 target_include_directories(lightning_gpu_algorithms PUBLIC ${CMAKE_CURRENT_SOURCE_DIR} )
